Introduce abstract IProjectLogger
